+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support If-None-Match header in `PUT /api/nodes/{id}`
 - Added `PUT|DELETE /api/overlays/{name}/file?path={path}`
 - Added `DELETE /api/overlays/{name}/file?path={path}`
+- Added `wwctl configure warewulfd`
 
 ### Changed
 
@@ -24,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fixed a bug when cloning an overlay to site when parent is missing
 - Fixed `wwctl upgrade nodes` to properly handle kernel argument lists. #1938
+- Improved netplan support. #1873
 
 ## v4.6.2, 2025-07-09